Database structure reshape
- creating the structure of the new mario_u package.
- the code is organzied as follow:
    - interface module: will containt the database.py (for managing
      the model database) and the model interface
    - model module: will containt the mathematical model including
      the creating of variables and the equations
    - tools module that will include some utils, input-output file
      handling and more importantly
      the constants.

constants.py: (in tools module)
this script contains a series of constants of the model. a customzied
class for buidling the constants is created called Constant inherited
from dict while the getattr and settattr are modified. How to use it?
Lets assume that we need to define the default values for matrices:

DEFAULTS = Constant(A=2,B=3)

to access the data one may use:
DEFAULTS['A']
or
DEFAULTS.A

io.py: (in tools module)
this script contains two main classes:

SetReader:
the class should be initialized with a main xlsx file or a dict of
pd.DataFrames(and the keys are named as sheets) path and has a
series of fucntions and methods to:
1. read the excel files
2. reshape every piece of data
3. extract intermediate data

InputBuider:
this class should be initialized with a DataFactory class and is in
charge of building all the model matrices including:
1. Endogenized matrices like G and J that should be filled automatically
2. Empty Exogenous matrices that will be printed out to be filled by
   user

Both classes will be used inside the DataFactory

database.py: (in interface module)
This module contains the DataFactory class that is the main interface
for the mario_u database handling. The class shoould be initialized with
a main excle file. This class will use the SetReader and InputBuilder
classes respectively for:
1. Reading the data and reshaping them
2. Printing out and reading the exogenous data

The class is equiped with different properties for easier handling of
data